requirements.txt
datasets
anthropic
anthropic[bedrock]
backoff
botocore
boto3
openai

#chess dgm
requests
numpy
ollama

# SWE-Bench
beautifulsoup4
chardet
docker
ghapi
GitPython
pre-commit
python-dotenv
rich
unidiff

# pytest
pytest
pytest-asyncio
async_timeout

#chess tools
python-chess
stockfish

